updated for geometry and number of objects loss calculation
ComputeLoss(con, exposureid, lossid, computecol='Cost', **kwargs)

The previous parameter computeonvalue=True has been replaced by computecol with default to 'Cost'. you can passs either of 'Cost','Population','Geometry','Count' in that input parameter and it will compute loss accrodingly. the population and cost columns must have been defined previously with costColumn and populColumn in metadata. for geometry and counts it is not required.  Geometry will give 1 as loss for all points, which in turn becomes same as the count.